+++ src/app/Notifications/Notifications.tsx
@@ -51,13 +51,6 @@ export interface Notification {
   timestamp?: number;
 }
 
-export interface NotificationDrawerCategory {
-  title: string;
-  isExpanded: boolean;
-  notifications: Notification[];
-  unreadCount: number;
-}
-
 export class Notifications {
 
   private readonly _notifications: Notification[] = [];
